\name{HyPhy-package}
\alias{HyPhy-package}
\alias{HyPhy}
\docType{package}
\title{Macroevolutionary phylogentic analysis of species trees and gene trees}
\description{A Bay Area high level phylogenetic analysis package mostly using the birth-death process. 
Analysis of species tree branching times and simulation of species trees 
under a number of different time variable birth-death processes.  
Analysis of gene tree species tree reconciliations and simulations of gene trees in species trees under the birth-death process.
}
\details{
\tabular{ll}{
Package: \tab HyPhy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-25\cr
License: \tab GPL-2\cr
}
}
\author{
Nathaniel Malachi Hallinan <nmhallinan@gmail.com>
}
